# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/venues": {
      "get": {
        "operationId": "VenuesController_findAll",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "List of venues"
          }
        },
        "summary": "List all venues",
        "tags": [
          "Venues"
        ]
      },
      "post": {
        "operationId": "VenuesController_create",
        "parameters": [
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVenueDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venue created"
          }
        },
        "summary": "Create a new venue",
        "tags": [
          "Venues"
        ]
      }
    },
    "/api/venues/{id}": {
      "get": {
        "operationId": "VenuesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Venue details"
          },
          "404": {
            "description": "Venue not found"
          }
        },
        "summary": "Get venue by ID",
        "tags": [
          "Venues"
        ]
      },
      "patch": {
        "operationId": "VenuesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVenueDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venue updated"
          }
        },
        "summary": "Update venue",
        "tags": [
          "Venues"
        ]
      },
      "delete": {
        "operationId": "VenuesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "User ID for user-specific operations",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Venue deleted"
          }
        },
        "summary": "Delete venue",
        "tags": [
          "Venues"
        ]
      }
    },
...
  },
  "info": {
    "title": "EMCP API",
    "description": "Event Management & Coordination Platform REST API. Role is passed via the x-user-role header. User ID is passed via the x-user-id header for user-specific operations.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateVenueDto": {
        "type": "object",
        "properties": {
          "venueName": {
            "type": "string",
            "example": "Grand Auditorium"
          },
          "location": {
            "type": "string",
            "example": "Main Campus, Building 4"
          },
          "capacity": {
            "type": "number",
            "example": 500
          },
          "status": {
            "type": "string",
            "example": "Active"
          }
        },
        "required": [
          "venueName",
          "capacity"
        ]
      },
      "UpdateVenueDto": {
        "type": "object",
        "properties": {
          "venueName": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  }
}